Update PluginList.vb
I am working on a new plugin for the pinball community and found that B2SServer does not send all LED "D" updates.  Specifically, values of "0" would not be sent for any LED other than position 0.  The value 0 is used by ROMs to "clear" or "blank" a digit.  The above suggestion solves the problem for "D" TableElementTypeChar while leaving the prior logic in place for all other types (which, presumably wasn't broken).

Please merge this change (or one with the same result) to correct this problem.